processing_requests/requirements.txt
flask
pymongo
kafka-python
